Rename 6.py to dictionary.py
It is necessary to create (not programmatically) a text file, where each line describes the subject and the availability of lectures, practical and laboratory classes on this subject, and their number. Each subject mustn't necessarily have all types of classes. Form a dictionary containing the name of the subject and the total number of lessons on it. Display dictionary on screen.