Update and rename 7.py to Profit.py
Create (non-programmatically) a text file in which each line should contain information about the company: name, ownership, revenue, costs.

It is necessary to read the file line by line, calculate the profit of each company, as well as the average profit. If the firm has incurred losses, do not include it in the calculation of average profit.
Next, implement the list. It should contain a dictionary with firms and their profits, as well as a dictionary with average profit. If the firm made a loss, also add it to the dictionary (with the meaning of the loss).

Save the final list as a json object to the appropriate file.