Rename 1.py to TrafficLight class .py
Create a TrafficLight class (traffic light) and define one color attribute (color) and a running method (start). Implement the attribute as private. As part of the method, implement the switching of the traffic light to the modes: red, yellow, green. The duration of the first state (red) is 7 seconds, the second (yellow) is 2 seconds, the third (green) is up to you. Switching between modes should only be done in the order shown (red, yellow, green). Test the example by creating an instance and calling the described method.