Rename 2.py to class Road.py
Implement the class Road (road), in which to define the attributes: length (length), width (width). The values of these attributes must be passed when an instance of the class is created. Attributes to make protected. Determine the method for calculating the mass of asphalt required to cover the entire roadway. Use the formula: length * width * mass of asphalt to cover one square meter of the road with asphalt, 1 cm thick * number cm of the thickness of the canvas. Check the operation of the method.