Update and rename 5.py to class Stationery.py
Implement the class Stationery. Define the title attribute (name) and the draw method (drawing) in it. The method prints the message “Starting rendering.” Create three child classes Pen (pen), Pencil (pencil), Handle (marker). In each of the classes, implement an override of the draw method. For each of the classes, the methods must display a unique message. Create class instances and check what the described method outputs for each instance.