Rename 1.py to class Matrix.py
Implement the class Matrix (matrix). Provide an overload of the class constructor (method __init__()) that must accept data (a list of lists) to form a matrix.

The next step is to implement an overload of the __str__() method to display the matrix in the usual way.
Next, implement an overload of the __add__() method to implement the operation of adding two objects of the Matrix class (two matrices). The result of the addition must be a new matrix.