Rename 2.py to production of clothing.py
Implement a project for calculating the total consumption of fabric for the production of clothing. The main essence (class) of this project is clothing, which can have a specific name. The clothing types in this project are coats and suits. These types of clothes have parameters: size (for a coat) and height (for a suit). These can be ordinary numbers: V and H, respectively.
To determine the fabric consumption for each type of clothing, use the formulas: for a coat (V / 6.5 + 0.5), for a suit (2 * H + 0.3). Test the performance of these methods on real data.
Implement a total count of tissue consumption. Check-in practices the knowledge gained in this lesson: implement abstract classes for the main classes of the project, check the work of the @property decorator in practice.