Rename 1.py to class Date.py
Implement the class "Date", the constructor function of which must accept a date in the form of a string of the "day-month-year" format. Within the class, implement two methods. The first one, with the @classmethod decorator, should extract the number, month, year and convert their type to the "Number" type. The second one, with the @staticmethod decorator, should validate the day, month, and year (for example, the month is from 1 to 12). Check the operation of the resulting structure on real data.