Update and rename 3.py to class Error.py
Create your own exception class that should check the contents of the list for only numbers. Check the operation of the exception on a real example. It is necessary to ask the user for data and fill in the list. The exception class must control the data types of the list items.