Update and rename 4-5-6.py to class StoreMashines.py
Start working on the Office Equipment Warehouse project. Create a class that describes the warehouse. As well as the "Office equipment" class, which will be the base for the successor classes. These classes are specific types of office equipment (printer, scanner, copier). In the base class, define the parameters that are common to the cast types. In descendant classes, implement parameters that are unique for each type of office equipment.
Develop methods responsible for the receipt of office equipment at the warehouse and transfer to a specific division of the company. To store data on the name and number of office equipment units, as well as other data, you can use any suitable structure, such as a dictionary.
 Implement a validation mechanism for user input. For example, you cannot use the string data type to specify the number of printers shipped to the warehouse.