Rename 7.py to class ComplexNumber.py
Implement the project "Operations with complex numbers". Create a class "Complex number", implement an overload of the methods of addition and multiplication of complex numbers. Test the project by creating instances of the class (complex numbers) and performing addition and multiplication of the generated instances. Check the correctness of the result.